requirements.txt
python-dotenv
pandas
sqlalchemy
dash
dash-bootstrap-components
plotly